Extracted inductive instances up into MonadCancel
